Added Featcher: Export calendar to CSV format for GoogleCalendar/iCal/Outlook
at the gui - calendar tab - there is now an export to csv button.
clicking it will prompt thr user to save the file on his system hard
drive. the app will add the .csv extention.

the csv file, will export only one week, starting from the begining of
the first semester of 2015 (the one that is coming).

in the future i will add a Held doc for exporting and importimg to the
users calendar.

cheers!